modifying navigation in menu ui and move user state view model in view model package
